poller/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
schedule>=1.2.0
tenacity>=8.2.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=4.9.0